Added code to method that runs on application startup (in this case the Main method) to create and configure a Unity dependency injection container.  We then use that container to create a BankAccount object.
NOTE: Unlike the previous code, we never explicitly tell Unity to create a BankAccount object.  Instead we tell it we want an object that implements the IBankAccount interface.  Unity uses its config file to determine what type of object to create.  If we wanted to use a different type of object we would modify Unity's config file; we would not need to recompile the application.

NOTE 2: At this stage the application will build and run but the BankAccount_Transfer unit test will not.  The unit test calls the old BankAccount constructor, which does not have any arguments, so no ITransferService object is passed into the BankAccount object.  The unit test will fail when the BankAccount.Transfer method attempts to call the ITransferService.SendFunds method.  We need to create a mock ITransferService object and pass it into the new BankAccount constructor.  That will be the next step.